First stage of putting guards in inline functions.
The eventual goal of this work is to make Ward unnecessary. We
accomplish this by putting guards in functions such as ADDR_OBJ()
and CONST_ADDR_OBJ() so that one cannot access objects without going
through a guard (or accessing the memory directly).

Conversely, we need unsafe versions of such functions (UNSAFE_ADDR_OBJ())
that forgo the use of guards so that internal mechanisms that intentionally
bypass them do not trigger errors.