kernel: use custom GC marking for T_FUNCTION
Not all slots of a T_FUNCTION bag are filled with bag refs, yet when
marking the slots of such a bag during garbage collection, we still used
MarkAllSubBags. This is usually no problem with GASMAN, as it detects if
a pointer isn't a master pointer, and can then simply ignore it. But
other garbage collections can't as easily verify master pointers. So
let's try to be accurate about what we mark as a bag and what not: skip
the first eight slots of every T_FUNCTION bag for marking (they contain
pointers to C functions), and also ensure that the rest of any
T_FUNCTION bag only contains bag refs.

Also fix a bug in saving/restoring operations, where the 'enabled' field
was stored as an UInt, even though it now is an Obj (though we currently
only store immediate ints in it, hence there was no functional problem).