Add new garbage collector based on Julia
Use `--with-gc=julia` to switch to this alternative garbage
collector. There is also a new `--with-julia=` option which
can be used to point GAP at the correct Julia version (which
currently still needs some patches to add APIs we need).

The main work is in the new file src/julia_gc.c